backend/app/api/routers/community.py
from fastapi import APIRouter, Depends
from sqlalchemy.orm import Session
from .database import get_db

router = APIRouter()

@router.get("/recommend")
def recommend_communities(user_id: int, db: Session = Depends(get_db)):
    # Logic to recommend communities based on user tags
    pass
